se completa todos los endpoints con sus requerimientos

+++ Models/HistoryResponse.cs
@@ -3,6 +3,6 @@
     public class HistoryResponse
     {
         public List<HistoryInfo> Details { get; set; }
-        public int AccountBalance { get; set; }
+        public decimal AccountBalance { get; set; }
     }
 }
